rename folder; code readme update
