+++ docs/zero_to_hero_guide/email_agent/requirements.txt
@@ -1,4 +1,3 @@
-
 google-auth==2.27.0
 google-auth-oauthlib==0.4.6
 google-auth-httplib2==0.1.0
